Really fixing the encoder class this time.
MA3 magnetic encoder uses analog signals. It's not meant to plug
straight into the roboRIO. Code can reference the encoder when it's
plugged into an analog breakout board off of a Talon SRX. It will then
use the talon's CAN bus to communicate back. Check this commit to see
notation on how to init and use the encoder.