Separate Intake, More Auton, CAM style shooter
Started framework for cam style shooter. Took out the previous piston
actuated shooter. Made a separate intake class, as it will probably be
used for overcoming defenses. Barely added to auton for low bar and
added support for pto on gear box for climbing. Accounting for
pneumatics for the shifting.