Fixed Some UDP Stuff
We were getting old data if we pressed the button too fast, so we
decided to add a checksum. After adding this, I noticed we were using
UDP.getData() rather than UDP.flush(), so I swapped it back. Hopefully
this makes the checksum unnecessary, but I left it in as another
safeguard.

The framework added for the checksum will allow us to add in a ycord in
the same way, if/when we chose to do so.